Move the REPL code to its own file
